+++ python/cli.py
@@ -21,44 +21,44 @@ _REGISTRY: dict[tuple[str, str], tuple[str, str]] = {
     ("lint", "gh-body-inline"): ("lint_gh_body_inline", "main"),
     ("lint", "codex-exec-auth"): ("lint_codex_exec_auth", "main"),
     ("lint", "skill-invocations"): ("lint_skill_invocations", "main"),
-    ("git", "commit"): ("git_cli", "commit_main"),
-    ("git", "stage"): ("git_cli", "stage_main"),
-    ("git", "amend-add"): ("git_cli", "amend_add_main"),
-    ("git", "current-branch"): ("git_cli", "current_branch_main"),
-    ("git", "branch-info"): ("git_cli", "branch_info_main"),
-    ("git", "conflict-files"): ("git_cli", "conflict_files_main"),
-    ("git", "rebase-abort"): ("git_cli", "rebase_abort_main"),
-    ("git", "rebase-skip"): ("git_cli", "rebase_skip_main"),
-    ("git", "checkout-ours"): ("git_cli", "checkout_ours_main"),
-    ("git", "show-stage"): ("git_cli", "show_stage_main"),
-    ("git", "sync-local-main"): ("git_cli", "sync_local_main_main"),
-    ("git", "clean-tree"): ("git_cli", "clean_tree_main"),
-    ("git", "snapshot-untracked"): ("git_cli", "snapshot_untracked_main"),
-    ("git", "count-commits"): ("git_cli", "count_commits_main"),
-    ("git", "check-main-sync"): ("git_cli", "check_main_sync_main"),
-    ("git", "check-remote-branch"): ("git_cli", "check_remote_branch_main"),
-    ("git", "check-phantom-dirty"): ("git_cli", "check_phantom_dirty_main"),
-    ("git", "phantom-probe"): ("git_cli", "phantom_probe_main"),
-    ("push", "branch"): ("push_cli", "branch_main"),
-    ("push", "force"): ("push_cli", "force_main"),
-    ("push", "rebase"): ("push_cli", "rebase_main"),
-    ("push", "checkpoint-probe"): ("push_cli", "checkpoint_probe_main"),
-    ("pr", "create-branch"): ("pr_cli", "create_branch_main"),
-    ("pr", "create"): ("pr_cli", "create_main"),
-    ("pr", "body-update"): ("pr_cli", "body_update_main"),
-    ("pr", "checks"): ("pr_cli", "checks_main"),
-    ("pr", "closes-issue"): ("pr_cli", "closes_issue_main"),
-    ("merge", "pr"): ("merge_cli", "pr_main"),
-    ("gh", "resolve-repo"): ("gh_cli", "resolve_repo_main"),
-    ("gh", "remote-repo"): ("gh_cli", "remote_repo_main"),
-    ("gh", "run-logs"): ("gh_cli", "run_logs_main"),
-    ("gh", "workflow-path"): ("gh_cli", "workflow_path_main"),
-    ("ci", "wait"): ("ci_cli", "wait_main"),
-    ("ci", "status"): ("ci_cli", "status_main"),
-    ("ci", "decide"): ("ci_cli", "decide_main"),
-    ("ci", "failed-jobs"): ("ci_cli", "failed_jobs_main"),
-    ("ci", "behind-count"): ("ci_cli", "behind_count_main"),
-    ("ci", "rerun-failed"): ("ci_cli", "rerun_failed_main"),
+    ("git", "commit"): ("git", "commit_main"),
+    ("git", "stage"): ("git", "stage_main"),
+    ("git", "amend-add"): ("git", "amend_add_main"),
+    ("git", "current-branch"): ("git", "current_branch_main"),
+    ("git", "branch-info"): ("git", "branch_info_main"),
+    ("git", "conflict-files"): ("git", "conflict_files_main"),
+    ("git", "rebase-abort"): ("git", "rebase_abort_main"),
+    ("git", "rebase-skip"): ("git", "rebase_skip_main"),
+    ("git", "checkout-ours"): ("git", "checkout_ours_main"),
+    ("git", "show-stage"): ("git", "show_stage_main"),
+    ("git", "sync-local-main"): ("git", "sync_local_main_main"),
+    ("git", "clean-tree"): ("git", "clean_tree_main"),
+    ("git", "snapshot-untracked"): ("git", "snapshot_untracked_main"),
+    ("git", "count-commits"): ("git", "count_commits_main"),
+    ("git", "check-main-sync"): ("git", "check_main_sync_main"),
+    ("git", "check-remote-branch"): ("git", "check_remote_branch_main"),
+    ("git", "check-phantom-dirty"): ("git", "check_phantom_dirty_main"),
+    ("git", "phantom-probe"): ("git", "phantom_probe_main"),
+    ("push", "branch"): ("push", "branch_main"),
+    ("push", "force"): ("push", "force_main"),
+    ("push", "rebase"): ("push", "rebase_main"),
+    ("push", "checkpoint-probe"): ("push", "checkpoint_probe_main"),
+    ("pr", "create-branch"): ("pr", "create_branch_main"),
+    ("pr", "create"): ("pr", "create_main"),
+    ("pr", "body-update"): ("pr", "body_update_main"),
+    ("pr", "checks"): ("pr", "checks_main"),
+    ("pr", "closes-issue"): ("pr", "closes_issue_main"),
+    ("merge", "pr"): ("merge", "pr_main"),
+    ("gh", "resolve-repo"): ("gh", "resolve_repo_main"),
+    ("gh", "remote-repo"): ("gh", "remote_repo_main"),
+    ("gh", "run-logs"): ("gh", "run_logs_main"),
+    ("gh", "workflow-path"): ("gh", "workflow_path_main"),
+    ("ci", "wait"): ("ci", "wait_main"),
+    ("ci", "status"): ("ci", "status_main"),
+    ("ci", "decide"): ("ci", "decide_main"),
+    ("ci", "failed-jobs"): ("ci", "failed_jobs_main"),
+    ("ci", "behind-count"): ("ci", "behind_count_main"),
+    ("ci", "rerun-failed"): ("ci", "rerun_failed_main"),
     ("session", "setup"): ("session_env", "setup_main"),
     ("session", "write-env"): ("session_env", "write_env_main"),
     ("session", "read-key"): ("session_env", "read_key_main"),
