+++ python/test_git.py
@@ -12,6 +12,8 @@ import pytest
 import git
 from errors import ShipError
 from proc import CommandResult
+import phantom
+from test_support import RecordingRunner
 
 
 @dataclass
@@ -666,3 +668,95 @@ def test_force_push_recovery_dirty_worktree() -> None:
     result = git.force_push_recovery(runner, branch="feat", remote="origin")
     assert not result.pushed
     assert result.status == "dirty_worktree"
+
+
+# CLI contract tests migrated from test_git_cli.py.
+def _ok(stdout: str = "", stderr: str = "") -> CommandResult:
+    return CommandResult(("cmd",), 0, stdout, stderr, 0.01)
+
+
+def _fail(stderr: str = "") -> CommandResult:
+    return CommandResult(("cmd",), 1, "", stderr, 0.01)
+
+
+def test_count_commits_missing_main_status_file(monkeypatch: pytest.MonkeyPatch, tmp_path: Path, capsys: pytest.CaptureFixture[str]) -> None:
+    runner = RecordingRunner(responses=[_fail(), _fail()])
+    monkeypatch.setattr(git, "proc", runner)
+    status_file = tmp_path / "status"
+    monkeypatch.setenv("COUNT_COMMITS_STATUS_FILE", str(status_file))
+    assert git.count_commits_main([]) == 0
+    assert capsys.readouterr().out.strip() == "0"
+    assert status_file.read_text(encoding="utf-8").strip() == "missing_main_ref"
+
+
+def test_branch_info_emits_detached_empty_branch(monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]) -> None:
+    runner = RecordingRunner(responses=[_ok("abc123\n"), _ok("\n")])
+    monkeypatch.setattr(git, "proc", runner)
+    assert git.branch_info_main([]) == 0
+    out = capsys.readouterr().out
+    assert "HEAD_SHA=abc123" in out
+    assert "CURRENT_BRANCH=" in out
+
+
+def test_clean_tree_fail_closed_probe_error(monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]) -> None:
+    runner = RecordingRunner(responses=[CommandResult(("git",), 128, "", "no repo\n", 0.01)])
+    monkeypatch.setattr(git, "proc", runner)
+    assert git.clean_tree_main(["--fail-closed"]) == 1
+    out = capsys.readouterr().out
+    assert "CLEAN=unknown" in out
+    assert "PROBE_ERROR=git exited 128" in out
+
+
+def test_check_phantom_dirty_clean_omits_optional_keys(monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]) -> None:
+    runner = RecordingRunner()
+    monkeypatch.setattr(git, "proc", runner)
+
+    def _clean_probe(*_args: object, **_kwargs: object) -> phantom.PhantomDirtyResult:
+        return phantom.PhantomDirtyResult(status="clean")
+
+    monkeypatch.setattr(phantom, "check_phantom_dirty", _clean_probe)
+    assert (
+        git.check_phantom_dirty_main(
+            ["--baseline", "/tmp/base.z", "--step", "s1", "--phantom-paths-dir", "/tmp/p"],
+        )
+        == 0
+    )
+    out = capsys.readouterr().out
+    assert "STATUS=clean" in out
+    assert "REASON=" not in out
+    assert "PHANTOM_COUNT=" not in out
+    assert "PHANTOM_PATHS_FILE=" not in out
+
+
+def test_check_phantom_dirty_parse_error_emits_unknown(monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]) -> None:
+    runner = RecordingRunner()
+    monkeypatch.setattr(git, "proc", runner)
+    assert git.check_phantom_dirty_main(["--unknown"]) == 0
+    out = capsys.readouterr().out
+    assert "STATUS=unknown" in out
+    assert "REASON=unknown-flag" in out
+
+
+def test_emit_kv_rejects_multiline_values() -> None:
+    with pytest.raises(ValueError, match="newline"):
+        git._emit_kv("ERROR", "line1\nline2")  # pyright: ignore[reportPrivateUsage]
+
+
+def test_snapshot_untracked_usage_does_not_create_output(tmp_path: Path) -> None:
+    output = tmp_path / "should-not-exist.z"
+    assert git.snapshot_untracked_main(["--unknown", str(output)]) == 0
+    assert not output.exists()
+
+
+def test_phantom_probe_clean_omits_optional_keys(monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]) -> None:
+    def _clean_probe(*_args: object, **_kwargs: object) -> phantom.PhantomProbeResult:
+        return phantom.PhantomProbeResult(
+            dirty=phantom.PhantomDirtyResult(status="clean"),
+        )
+
+    monkeypatch.setattr(phantom, "probe_with_warn", _clean_probe)
+    assert git.phantom_probe_main(["--step", "s1"]) == 0
+    out = capsys.readouterr().out
+    assert "PHANTOM_STATUS=clean" in out
+    assert "PHANTOM_COUNT=" not in out
+    assert "PHANTOM_PATHS_FILE=" not in out
