--- python/test_push_cli.py
@@ -1,67 +1,0 @@
-# pyright: reportUnknownParameterType=false, reportMissingParameterType=false, reportUnknownMemberType=false, reportUnknownVariableType=false, reportUnknownArgumentType=false, reportUnknownLambdaType=false, reportUnusedCallResult=false
-"""CLI contract tests for push_cli."""
-
-from __future__ import annotations
-
-from dataclasses import dataclass
-
-from proc import CommandResult
-from test_support import RecordingRunner
-
-import push_cli
-
-
-def _res(rc: int = 0, stdout: str = "", stderr: str = "") -> CommandResult:
-    return CommandResult(("cmd",), rc, stdout, stderr, 0.01)
-
-
-def test_force_status_map_dirty_worktree(monkeypatch, capsys):
-    runner = RecordingRunner(responses=[_res(stdout="feature\n"), _res(stdout=" M file\n")])
-    monkeypatch.setattr(push_cli, "proc", runner)
-    assert push_cli.force_main([]) == 1
-    out = capsys.readouterr().out
-    assert "PUSHED=false" in out
-    assert "STATUS=dirty_worktree" in out
-
-
-def test_checkpoint_probe_emits_rebase_outcome_on_skip(monkeypatch, capsys):
-    @dataclass
-    class _RebaseResult:
-        exit_code: int = 0
-        skipped_already_fresh: bool = True
-        skipped_already_pushed: bool = False
-        conflict_files: str = ""
-        rebase_error: str = ""
-
-    @dataclass
-    class _ProbeResult:
-        dirty: object
-        append_warn_error: str = ""
-
-    @dataclass
-    class _Dirty:
-        status: str = "clean"
-        reason: str = ""
-        count: int = 0
-        paths_file: str = ""
-
-    monkeypatch.setattr(push_cli.rebase, "rebase_push", lambda *_a, **_k: _RebaseResult())
-    monkeypatch.setattr(
-        push_cli.phantom,
-        "probe_with_warn",
-        lambda *_a, **_k: _ProbeResult(dirty=_Dirty()),
-    )
-    assert push_cli.checkpoint_probe_main(["1.r", "plan"]) == 0
-    out = capsys.readouterr().out
-    assert "REBASE_OUTCOME=skipped" in out
-    assert "SKIPPED_ALREADY_FRESH=true" in out
-    assert "PHANTOM_STATUS=clean" in out
-
-
-def test_branch_push_dedupes_stderr(monkeypatch, capsys):
-    runner = RecordingRunner(responses=[_res(stdout="feature\n"), _res(stdout="feature\n"), _res(1, stderr="nope\n"), _res(stdout="feature\n"), _res(1, stderr="nope\n"), _res(stdout="feature\n"), _res(1, stderr="nope\n")])
-    monkeypatch.setattr(push_cli, "proc", runner)
-    assert push_cli.branch_main([]) == 1
-    captured = capsys.readouterr()
-    assert "BRANCH=feature" in captured.out
-    assert "(repeated 3 times)" in captured.err
